src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/gc-shadow-expected.txt

CONSOLE ERROR: line 20: Uncaught TypeError: p.createShadowRoot is not a function
This tests that the JavaScript wrapper objects of shadow DOM objects
are not prematurely garbage collected.


FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
